pyproject.toml
[project]
name = "to-do-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.118.0",
    "httpx>=0.28.1",
    "pre-commit>=4.3.0",
    "pydantic-settings>=2.11.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn[standard]>=0.37.0",
]
